@@ -1,6 +1,5 @@
 import path = require('path');
-
-import fs = require('fs-promise');
+import * as fs from './readDirAsync';
 import Promise = require('bluebird');
 import * as vscode from 'vscode';
 
@@ -13,7 +12,7 @@ export function promiseList() : Promise<resourceBundle[]> {
     let resourceBundlesPath = buildResourceBundlesPath();
 
     let listResourceBundlesInWorkspace = fs.readdir(resourceBundlesPath)
-        .then(resourceBundlesFromFileList, console.error);
+        .then(resourceBundlesFromFileList);
     
     return listResourceBundlesInWorkspace;
 }
